Main

+++ src/users/routers.py
@@ -1,9 +1,11 @@
 import bcrypt
+from fastapi import *
 from fastapi.responses import *
 
 from src.config import secret_key
+
+from src.users.service import *
 from src.users.dependencies import *
-from src.users.schemas import *
 from src.users.crud import *
 
 router = APIRouter()
@@ -11,13 +13,15 @@ tags_router = ["api-user"]
 
 @router.get("/api/user/update-tokens", tags=tags_router)
 async def update_tokens(response: Response, refresh: str = Cookie()):
-    try:
-        data_refresh = jwt.decode(refresh, secret_key, algorithms="HS256")
+        data_refresh = decode_token(token=refresh, secret_key=secret_key)
 
-        return create_tokens(response, UserDataLogin(username=data_refresh["username"], password=data_refresh["password"]))
-    except:
-        response.delete_cookie("refresh")
-        return HTTPException(401, "Error refresh token")
+        if data_refresh is not None:
+            tokens = create_tokens(username=data_refresh["username"], password=data_refresh["password"], secret_key=secret_key)
+            response.set_cookie("refresh", tokens["refresh"], httponly=True)
+            return tokens
+        else:
+            response.delete_cookie("refresh")
+            raise HTTPException(401, "Error refresh token")
 
 @router.post("/api/user/login", tags=tags_router)
 async def login(response: Response, user_data: UserDataLogin = Body()):
@@ -25,11 +29,14 @@ async def login(response: Response, user_data: UserDataLogin = Body()):
 
     if user is not None:
         if bcrypt.checkpw(user_data.password.encode(), user.password.encode()) and user_data.username == user.username:
-            return create_tokens(response, user)
+            tokens = create_tokens(username=user.username, password=user.password,
+                                   secret_key=secret_key)
+            response.set_cookie("refresh", tokens["refresh"], httponly=True)
+            return tokens
         else:
-            return HTTPException(409, "Incorrect data")
+            raise HTTPException(409, "Incorrect data")
     else:
-        return HTTPException(401, "unauthorized")
+        raise HTTPException(401, "unauthorized")
 
 
 @router.post("/api/user/regist", tags=tags_router)
@@ -37,10 +44,14 @@ async def regist(response: Response, user_data: UserDataRegist = Body()):
     user_data.password = hash_pw(user_data.password)
     
     if get_user(username=user_data.username) is not None:
-        return HTTPException(409, "The use already exists")
+        raise HTTPException(409, "The use already exists")
     
     add_user(user_data)
-    return create_tokens(response, user_data)
+
+    tokens = create_tokens(username=user_data.username, password=user_data.password,
+                           secret_key=secret_key)
+    response.set_cookie("refresh", tokens["refresh"], httponly=True)
+    return tokens
 
 @router.post("/api/user/logout", tags=tags_router)
 async def logout(response: Response):
